# faster output of comments