#### Before
```
$ nvs use node
PATH = node/9.3.0/x64

$ uglifyjs ember.prod.js --comments all --timings | node
- parse: 0.421s
- rename: 0.000s
- compress: 0.000s
- scope: 0.000s
- mangle: 0.000s
- properties: 0.000s
- output: 3.256s
- total: 3.677s

$ nvs use chakracore
PATH -= chakracore/8.6.0/x64

$ uglifyjs ember.prod.js --comments all --timings | node
- parse: 0.621s
- rename: 0.000s
- compress: 0.000s
- scope: 0.000s
- mangle: 0.000s
- properties: 0.000s
- output: 22.834s
- total: 23.455s
```

#### After
```
$ nvs use node
PATH = node/9.3.0/x64

$ uglifyjs ember.prod.js --comments all --timings | node
- parse: 0.421s
- rename: 0.000s
- compress: 0.000s
- scope: 0.000s
- mangle: 0.000s
- properties: 0.000s
- output: 1.907s
- total: 2.328s

$ nvs use chakracore
PATH -= chakracore/8.6.0/x64

$ uglifyjs ember.prod.js --comments all --timings | node
- parse: 0.628s
- rename: 0.000s
- compress: 0.000s
- scope: 0.000s
- mangle: 0.000s
- properties: 0.000s
- output: 1.557s
- total: 2.185s
```