@kzc 2018-01-17
Good idea to optimize the common case.

Looks like `chakracore` JIT lacks compiled regex caching.