@alexlamsl 2018-01-17
I was actually looking at possible `setBlocking()` issue with Node.js 4 on Windows, and was trying to boost the output by using `--comments all` on concatenated copies of `ember.prod.js` - that went well...

Only discover the worse slow-down on `node-chakracore` when I tried to verify this PR doesn't cause regression on that platform.