@kzc 2018-01-17
OT - mentioned this a while back but couldn't find the link... dependent packages by version:

https://libraries.io/npm/uglify-js/usage
https://libraries.io/npm/uglify-es/usage