@alexlamsl 2018-01-17
Ah, so they are counting repositories instead of download usage - that's one way of looking at it I guess.