@kzc 2018-01-17
Nevermind that the downloads per repo would vary wildly. It just gives a rough idea of the version distribution in use.