@alexlamsl 2018-01-18
I'm more worried about those (tutorial) repositories where users are instructed to fork in order to use it.